Comments fix
